Interface clean up (WIP)
