+++ src/main/java/com/github/dakusui/actionunit/actions/cmd/Commander.java
@@ -11,40 +11,6 @@ import static com.github.dakusui.processstreamer.core.process.ProcessStreamer.Ch
 import static java.util.Objects.requireNonNull;
 
 public interface Commander<C extends Commander<?>> {
-  class RetryOption {
-    public static RetryOption timeoutInSeconds(long timeoutInSeconds) {
-      return new RetryOption(
-          timeoutInSeconds,
-          TimeUnit.SECONDS,
-          Throwable.class,
-          0,
-          TimeUnit.SECONDS,
-          0);
-    }
-
-    final long timeoutDuration;
-    final TimeUnit timeoutTimeUnit;
-    final Class<? extends Throwable> retryOn;
-    final long retryInterval;
-    final TimeUnit retryIntervalTimeUnit;
-    final int retries;
-
-
-    RetryOption(
-        long timeoutDuration,
-        TimeUnit timeoutTimeUnit,
-        Class<? extends Throwable> retryOn,
-        long retryInterval,
-        TimeUnit retryIntervalTimeUnit,
-        int retries) {
-      this.timeoutDuration = timeoutDuration;
-      this.timeoutTimeUnit = requireNonNull(timeoutTimeUnit);
-      this.retryOn = requireNonNull(retryOn);
-      this.retryInterval = retryInterval;
-      this.retryIntervalTimeUnit = requireNonNull(retryIntervalTimeUnit);
-      this.retries = retries;
-    }
-  }
 
   C env(String varname, String varvalue);
